fixed cursor info display to display correct values from spatial profile stream
