+++ src/models/CursorInfo.ts
@@ -5,5 +5,4 @@ export interface CursorInfo {
     posImageSpace: Point2D;
     posWCS: Point2D;
     infoWCS: { x: string, y: string };
-    value: number;
 }
